+++ app/controllers/NewsController.php
@@ -19,6 +19,17 @@ class NewsController extends BaseController
      */
     protected $layout = 'layouts.master';
 
+    public function __construct()
+    {
+        if (Auth::check()) {
+            $userId = Auth::id();
+            $user = User::findOrFail($userId);
+            $role = $user->role_id;
+            if($role == 1) {
+                $this->layout = 'layouts.masterlogin';
+            }
+        }
+    }
     /**
      * Index news action.
      *
@@ -28,11 +39,24 @@ class NewsController extends BaseController
     {
         $tree = Tree::findOrFail(4);
         if ( $tree->active == 1) {
+            if (Auth::check()) {
+                $userId = Auth::id();
+                $user = User::findOrFail($userId);
+                $role = $user->role_id;
+                if($role == 1) {
+                    $actions = 1;
+                } else {
+                    $actions = 0;
+                }
+            } else {
+                $actions = 0;
+            }
             $news = News::paginate(6);
             $news_lead = Tree::findOrFail(4);
             $this->layout->content = View::make('news.index', array(
                 'news' => $news,
                 'news_lead' => $news_lead,
+                'actions' => $actions,
             ));
         } else {
             return Redirect::route('homepage');
@@ -49,11 +73,24 @@ class NewsController extends BaseController
     {
         $tree = Tree::findOrFail(4);
         if ( $tree->active == 1) {
+            if (Auth::check()) {
+                $userId = Auth::id();
+                $user = User::findOrFail($userId);
+                $role = $user->role_id;
+                if($role == 1) {
+                    $actions = 1;
+                } else {
+                    $actions = 0;
+                }
+            } else {
+                $actions = 0;
+            }
             $news = DB::table('news')->where('course_id', "=", $id)->paginate(6);
             $news_lead = Tree::findOrFail(4);
             $this->layout->content = View::make('news.index', array(
                 'news' => $news,
-                'news_lead' => $news_lead
+                'news_lead' => $news_lead,
+                'actions' => $actions,
             ));
         } else {
             return Redirect::route('homepage');
@@ -70,16 +107,28 @@ class NewsController extends BaseController
     {
         $tree = Tree::findOrFail(4);
         if ( $tree->active == 1) {
-            $news = News::findOrFail($id);
-            $courses = Course::all()->lists('name', 'id');
-            $tags = Tag::all()->lists('name','id');
-            $news_tags = $news->tags()->lists('tag_id');
-            $this->layout->content = View::make('news.edit', array(
-                'news' => $news,
-                'courses' => $courses,
-                'tags' => $tags,
-                'news_tags' => $news_tags,
-            ));
+            if (Auth::check()) {
+                $userId = Auth::id();
+                $user = User::findOrFail($userId);
+                $role = $user->role_id;
+                if($role == 1) {
+
+                    $news = News::findOrFail($id);
+                    $courses = Course::all()->lists('name', 'id');
+                    $tags = Tag::all()->lists('name','id');
+                    $news_tags = $news->tags()->lists('tag_id');
+                    $this->layout->content = View::make('news.edit', array(
+                        'news' => $news,
+                        'courses' => $courses,
+                        'tags' => $tags,
+                        'news_tags' => $news_tags,
+                    ));
+                } else {
+                    return Redirect::route('news.view', $id);
+                }
+            } else {
+                return Redirect::route('news.view', $id);
+            }
         } else {
             return Redirect::route('homepage');
         }
@@ -94,12 +143,23 @@ class NewsController extends BaseController
     {
         $tree = Tree::findOrFail(4);
         if ( $tree->active == 1) {
-            $courses = Course::all()->lists('name', 'id');
-            $tags = Tag::all()->lists('name', 'id');
-            $this->layout->content = View::make('news.new', array(
-                'courses' => $courses,
-                'tags' => $tags
-            ));
+            if (Auth::check()) {
+                $userId = Auth::id();
+                $user = User::findOrFail($userId);
+                $role = $user->role_id;
+                if($role == 1) {
+                    $courses = Course::all()->lists('name', 'id');
+                    $tags = Tag::all()->lists('name', 'id');
+                    $this->layout->content = View::make('news.new', array(
+                        'courses' => $courses,
+                        'tags' => $tags
+                    ));
+                } else {
+                    return Redirect::route('news.index');
+                }
+            } else {
+                return Redirect::route('news.index');
+            }
         } else {
             return Redirect::route('homepage');
         }
@@ -115,31 +175,42 @@ class NewsController extends BaseController
     {
         $tree = Tree::findOrFail(4);
         if ( $tree->active == 1) {
-            $news = News::findOrFail($id);
-            $news->title = Input::get('title');
-            $news->lead = Input::get('lead');
-            $news->date = Input::get('date');
-            $news->content = Input::get('content');
-            $news->course_id = Input::get('courses');
+            if (Auth::check()) {
+                $userId = Auth::id();
+                $user = User::findOrFail($userId);
+                $role = $user->role_id;
+                if($role == 1) {
+                    $news = News::findOrFail($id);
+                    $news->title = Input::get('title');
+                    $news->lead = Input::get('lead');
+                    $news->date = Input::get('date');
+                    $news->content = Input::get('content');
+                    $news->course_id = Input::get('courses');
 
-            $rules = $news->rules();
-            $validator = Validator::make(Input::all(), $rules);
+                    $rules = $news->rules();
+                    $validator = Validator::make(Input::all(), $rules);
 
-            if ($validator->fails())
-            {
-                return Redirect::route('news.edit', $news->id)->withErrors($validator);
+                    if ($validator->fails())
+                    {
+                        return Redirect::route('news.edit', $news->id)->withErrors($validator);
+                    } else {
+                        $news->save();
+                        if(Input::get('tags') != NULL){
+                            $news->tags()->sync(Input::get('tags'));
+                        }
+                        $tree_students = Tree::findOrFail(2);
+                        $action = 'update';
+                        if( $tree_students->active == 1) {
+                            $news->sendMail($news->course_id, $action);
+                        }
+                        Session::flash('message', Lang::get('app.news-updated'));
+                        return Redirect::route('news.view', $news->id);
+                    }
+                } else {
+                    return Redirect::route('news.view', $id);
+                }
             } else {
-                $news->save();
-                if(Input::get('tags') != NULL){
-                    $news->tags()->sync(Input::get('tags'));
-                }
-                $tree_students = Tree::findOrFail(2);
-                $action = 'update';
-                if( $tree_students->active == 1) {
-                    $news->sendMail($news->course_id, $action);
-                }
-                Session::flash('message', Lang::get('app.news-updated'));
-                return Redirect::route('news.view', $news->id);
+                return Redirect::route('news.view', $id);
             }
         } else {
             Session::flash('message', Lang::get('common.no-such-site'));
@@ -156,35 +227,46 @@ class NewsController extends BaseController
     {
         $tree = Tree::findOrFail(4);
         if ( $tree->active == 1) {
-            $news = new News();
-            $news->title = Input::get('title');
-            $news->lead = Input::get('lead');
-            $news->date = Input::get('date');
-            $news->content = Input::get('content');
-            $news->course_id = Input::get('courses');
-            $news->owner_id = 1;
-            $news->owner_role_id = 1;
+            if (Auth::check()) {
+                $userId = Auth::id();
+                $user = User::findOrFail($userId);
+                $role = $user->role_id;
+                if($role == 1) {
+                    $news = new News();
+                    $news->title = Input::get('title');
+                    $news->lead = Input::get('lead');
+                    $news->date = Input::get('date');
+                    $news->content = Input::get('content');
+                    $news->course_id = Input::get('courses');
+                    $news->owner_id = 1;
+                    $news->owner_role_id = 1;
 
-            $rules = $news->rules();
-            $validator = Validator::make(Input::all(), $rules);
-            
-            if ($validator->fails())
-            {
-                return Redirect::route('news.new')
-                    ->withErrors($validator)
-                    ->withInput();
+                    $rules = $news->rules();
+                    $validator = Validator::make(Input::all(), $rules);
+
+                    if ($validator->fails())
+                    {
+                        return Redirect::route('news.new')
+                            ->withErrors($validator)
+                            ->withInput();
+                    } else {
+                        $news->save();
+                        if(Input::get('tags') != NULL){
+                            $news->tags()->sync(Input::get('tags'));
+                        }
+                        $tree_students = Tree::findOrFail(2);
+                        $action = 'update';
+                        if( $tree_students->active == 1) {
+                            $news->sendMail($news->course_id, $action);
+                        }
+                        Session::flash('message', Lang::get('app.news-created'));
+                        return Redirect::route('news.view', $news->id);
+                    }
+                } else {
+                    return Redirect::route('news.index');
+                }
             } else {
-                $news->save();
-                if(Input::get('tags') != NULL){
-                    $news->tags()->sync(Input::get('tags'));
-                }
-                $tree_students = Tree::findOrFail(2);
-                $action = 'update';
-                if( $tree_students->active == 1) {
-                    $news->sendMail($news->course_id, $action);
-                }
-                Session::flash('message', Lang::get('app.news-created'));
-                return Redirect::route('news.view', $news->id);
+                return Redirect::route('news.index');
             }
         } else {
             Session::flash('message', Lang::get('common.no-such-site'));
@@ -202,9 +284,22 @@ class NewsController extends BaseController
     {
         $tree = Tree::findOrFail(4);
         if ( $tree->active == 1) {
+            if (Auth::check()) {
+                $userId = Auth::id();
+                $user = User::findOrFail($userId);
+                $role = $user->role_id;
+                if($role == 1) {
+                    $actions = 1;
+                } else {
+                    $actions = 0;
+                }
+            } else {
+                $actions = 0;
+            }
             $news = News::findOrFail($id);
             $this->layout->content = View::make('news.view',array(
                 'news' => $news,
+                'actions' => $actions,
             ));
         } else {
             return Redirect::route('homepage');
@@ -221,10 +316,21 @@ class NewsController extends BaseController
     {
         $tree = Tree::findOrFail(4);
         if ( $tree->active == 1) {
-            $news = News::findOrFail($id);
-            $this->layout->content = View::make('news.delete', array(
-                'news' => $news,
-            ));
+            if (Auth::check()) {
+                $userId = Auth::id();
+                $user = User::findOrFail($userId);
+                $role = $user->role_id;
+                if($role == 1) {
+                    $news = News::findOrFail($id);
+                    $this->layout->content = View::make('news.delete', array(
+                        'news' => $news,
+                    ));
+                } else {
+                    return Redirect::route('news.view', $id);
+                }
+            } else {
+                return Redirect::route('news.view', $id);
+            }
         } else {
             return Redirect::route('homepage');
         }
@@ -240,11 +346,22 @@ class NewsController extends BaseController
     {
         $tree = Tree::findOrFail(4);
         if ( $tree->active == 1) {
-            $news = News::findOrFail($id);
-            $news->tags()->detach();
-            $news->delete();
-            Session::flash('message', Lang::get('app.news-destroyed'));
-            return Redirect::route('news.index');
+            if (Auth::check()) {
+                $userId = Auth::id();
+                $user = User::findOrFail($userId);
+                $role = $user->role_id;
+                if($role == 1) {
+                    $news = News::findOrFail($id);
+                    $news->tags()->detach();
+                    $news->delete();
+                    Session::flash('message', Lang::get('app.news-destroyed'));
+                    return Redirect::route('news.index');
+                } else {
+                    return Redirect::route('news.view', $id);
+                }
+            } else {
+                return Redirect::route('news.view', $id);
+            }
         } else {
             Session::flash('message', Lang::get('common.no-such-site'));
             return Redirect::route('homepage');
